pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webprobe"
version = "0.2.0"
description = "Generic site state-graph auditor — map, capture, analyze, report, explore"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "playwright>=1.40",
    "networkx>=3.0",
    "aiohttp>=3.9",
    "jinja2>=3.1",
]

[project.optional-dependencies]
explore = [
    "anthropic>=0.40",
    "transmogrifier",
    "Pillow>=10.0",
]
openai = [
    "openai>=1.0",
]
gemini = [
    "google-genai>=1.0",
]
apprentice = [
    "apprentice",
]
all-llm = [
    "webprobe[explore,openai,gemini,apprentice]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
]

[project.scripts]
webprobe = "webprobe.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/webprobe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
